+++ backend/bin/run-jobs2.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-DIR=$(dirname "$(realpath "$0")")
-
-"${DIR}"/console update-chars --log --hide-details
-"${DIR}"/console update-player-groups --log --hide-details